requirements.txt
pyautogen==0.1.13
requests~=2.31.0
streamlit==1.26.0
openai==0.28.0
toml~=0.10.2
pandas==2.1.1
tabulate==0.9.0
python-dotenv==1.0.0
httpx==0.25.0
asyncio
pytest
